IGNITE-26012 Base test extracted
